refactor: Game 클래스를 controller 패키지 안으로 이동
